Module bamboo_rs_core::entry

Re-exports

pub use decode::decode;

pub use publish::publish;

pub use verify::verify;

pub use verify::verify_batch;

Modules

decode

encode

publish

verify

Structs

Entry

Constants

MAX_ENTRY_SIZE
This is useful if you need to know at compile time how big an entry can get.

MAX_ENTRY_SIZE_

MAX_VARU64_SIZE

PUBLIC_KEY_LENGTH
The length of an ed25519 PublicKey, in bytes.

TAG_BYTE_LENGTH

Functions

into_owned

is_lipmaa_required